Update Stack.java
Useful method for the move code, should simplify everyone elses work. Simply tell splitStackContentsForMove how many of its pieces are leaving the stack, and it will return a list you can give the captureToken() method of another stack.  This is practically the entire operation of moving from on tile to another., other than the class that knows and can communicate to both stacks.